requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-docx==0.8.11
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0